pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "bookin"
version = "0.1.0"
description = "Ebook folder watcher and organizer using Calibre"
requires-python = ">=3.11"
dependencies = [
    "watchdog>=4.0",
    "pyyaml>=6.0",
    "click>=8.1",
    "rich>=13.0",
]

[project.scripts]
bookin = "bookin.cli:main"

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-mock>=3.12",
    "ruff>=0.4",
    "mypy>=1.10",
    "types-pyyaml>=6.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/bookin"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "SIM", # flake8-simplify
]
ignore = [
    "E501", # line too long (handled by formatter)
]

[tool.ruff.lint.isort]
known-first-party = ["bookin"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_ignores = true
warn_return_any = true

[[tool.mypy.overrides]]
module = "watchdog.*"
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
